+++ common/Tryte.h
