+++ common/util.h
