+++ res/ascii_tilemap.bmp
