+++ triumph/triumph.vcxproj.user
@@ -1,11 +1,11 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
-    <LocalDebuggerCommandArguments>-d $(ProjectDir)\res\helloworld.td</LocalDebuggerCommandArguments>
+    <LocalDebuggerCommandArguments>-d $(SolutionDir)\res\helloworld.td</LocalDebuggerCommandArguments>
     <DebuggerFlavor>WindowsLocalDebugger</DebuggerFlavor>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
-    <LocalDebuggerCommandArguments>-d $(ProjectDir)\res\helloworld.td</LocalDebuggerCommandArguments>
+    <LocalDebuggerCommandArguments>-d $(SolutionDir)\res\helloworld.td</LocalDebuggerCommandArguments>
     <DebuggerFlavor>WindowsLocalDebugger</DebuggerFlavor>
   </PropertyGroup>
 </Project>
